Construct the ADDRESS_STRUCTURE tibble

This function constructs a tibble representation of the ADDRESS_STRUCTURE from the GEDCOM 5.5.5 specification.

ADDRESS_STRUCTURE(
  local_address_lines = character(),
  address_city = character(),
  address_state = character(),
  address_postal_code = character(),
  address_country = character(),
  phone_number = character(),
  address_email = character(),
  address_fax = character(),
  address_web_page = character()
)

Arguments

local_address_lines
These address lines contain the address lines before the town/city. This can be a vector with up to 3 elements.

address_city
The name of the city/town used in the address. Isolated for sorting or indexing.

address_state
The name of the province, state or similar country subdivision used in the address. Isolated for sorting or indexing.

address_postal_code
The ZIP or postal code used by the various localities in handling of mail. Isolated for sorting or indexing.

address_country
The name of the country that pertains to the associated address. Isolated by some systems for sorting or indexing. Used in most cases to facilitate automatic sorting of mail.

phone_number
A phone number.

address_email
An electronic address that can be used for contact such as an email address.

address_fax
A FAX telephone number appropriate for sending data facsimiles.

address_web_page
The world wide web page address.

Value

A tidy tibble containing the ADDRESS_STRUCTURE part of a GEDCOM file.