Construct the EVENT_DETAIL tibble

This function constructs a tibble representation of the EVENT_DETAIL from the GEDCOM 5.5.5 specification.

EVENT_DETAIL(
  event_or_fact_classification = character(),
  date = date_value(),
  place = PLACE_STRUCTURE(character()),
  address = ADDRESS_STRUCTURE(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  notes = list(),
  source_citations = list(),
  multimedia_links = list()
)

Arguments

event_or_fact_classification
A descriptive word or phrase used to further classify the parent event or attribute tag. This should be used whenever either of the generic EVEN or FACT tags are used. The value of this primitive is responsible for classifying the generic event or fact being cited.

date
A date_value() object giving the date of the event.

place
A PLACE_STRUCTURE() object giving the location of the event.

address
An ADDRESS_STRUCTURE() object giving the address of the event.

responsible_agency
The organization, institution, corporation, person, or other entity that has responsibility for the associated context. For example, an employer of a person of an associated occupation, or a church that administered rites or events, or an organization responsible for creating and/or archiving records.

religious_affiliation
A name of the religion with which this person, event, or record was affiliated.

cause_of_event
Used in special cases to record the reasons which precipitated an event. Normally this will be used subordinate to a death event to show cause of death, such as might be listed on a death certificate.

notes
A list of NOTE_STRUCTURE() objects.

source_citations
A list of SOURCE_CITATION() objects.

multimedia_links
A list of MULTIMEDIA_LINK() objects

Value

A tidy tibble containing the EVENT_DETAIL part of a GEDCOM file.