Construct the LINEAGE_LINKED_HEADER_EXTENSION tibble

This function constructs a tibble representation of the LINEAGE_LINKED_HEADER_EXTENSION from the GEDCOM 5.5.5 specification.

LINEAGE_LINKED_HEADER_EXTENSION(
  system_id = "tidygedcom",
  product_version_number = utils::packageVersion("tidygedcom"),
  name_of_product = "tidygedcom",
  name_of_business = "Jamie Lendrum",
  business_address = ADDRESS_STRUCTURE(address_email = "jalendrum@gmail.com",
    address_web_page = "https://jl5000.github.io/tidygedcom/"),
  name_of_source_data = character(),
  publication_date = date_exact(),
  copyright_source_data = character(),
  receiving_system_name = character(),
  file_creation_date = date_exact(),
  file_creation_time = character(),
  language_of_text = character(),
  xref_subm = character(),
  gedcom_file_name = character(),
  copyright_gedcom_file = character(),
  gedcom_content_description = character()
)

Arguments

system_id
A system identification name. This name must be unique for each system (product), different from any other system. The name may include spaces, and is not restricted to ASCII characters.

product_version_number
The version of the product that created the GEDCOM file. It is defined and changed by the creators of the product.

name_of_product
The name of the software product that produced this file.

name_of_business
Name of the business, corporation, or person that produced or commissioned the product.

business_address
An ADDRESS_STRUCTURE() object giving the address of the business.

name_of_source_data
The name of the electronic data source that was used to obtain the data in this file. For example, the data may have been obtained from a CD-ROM disc that was named "U.S. 1880 CENSUS CD-ROM vol. 13."

publication_date
A date_exact() object giving the date this source was published or created.

copyright_source_data
A copyright statement required by the owner of data from which this information was downloaded.

receiving_system_name
The name of the system expected to process the GEDCOM file.

file_creation_date
A date_exact() object giving the date that this file was created.

file_creation_time
The time of a specific event.

language_of_text
The human language in which the data in the file is normally read or written. It is used primarily by programs to select language-specific sorting sequences and phonetic name matching algorithms.

xref_subm
An xref ID of a submitter record.

gedcom_file_name
The name of the GEDCOM file. A GEDCOM file name should use the format basename.ext, and use the file extension .GED (or .ged).

copyright_gedcom_file
A copyright statement needed to protect the copyrights of the submitter of this GEDCOM file.

gedcom_content_description
A note that a user enters to describe the contents of the lineage-linked file in terms of "ancestors or descendants of" so that the person receiving the data knows what genealogical information the file contains.

Value

A tidy tibble containing the HEADER part of a GEDCOM file.