Construct the SOURCE_RECORD tibble

This function constructs a tibble representation of the SOURCE_RECORD from the GEDCOM 5.5.5 specification.

SOURCE_RECORD(
  xref_sour,
  events_recorded = character(),
  date_period_covered = date_period(),
  source_jurisdiction_place = character(),
  responsible_agency = character(),
  data_notes = list(),
  source_originator = character(),
  source_descriptive_title = character(),
  source_filed_by_entry = character(),
  source_publication_facts = character(),
  text_from_source = character(),
  source_repository_citations = list(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  date_changed = CHANGE_DATE(),
  notes = list(),
  multimedia_links = list()
)

Arguments

xref_sour
An xref ID of a Source record.

events_recorded
An enumeration of the different kinds of events that were recorded in a particular source. Each enumeration is separated by a comma. Such as a parish register of births, deaths, and marriages would be BIRT, DEAT, MARR.

date_period_covered
A date_period() object associated with the period covered by the course.

source_jurisdiction_place
The name of the lowest jurisdiction that encompasses all lower-level places named in this source.

responsible_agency
The organization, institution, corporation, person, or other entity that has responsibility for the associated context. For example, an employer of a person of an associated occupation, or a church that administered rites or events, or an organization responsible for creating and/or archiving records.

data_notes
A list of NOTE_STRUCTURE() objects associated with the data in this source.

source_originator
The person, agency, or entity who created the record. For a published work, this could be the author, compiler, transcriber, abstractor, or editor. For an unpublished source, this may be an individual, a government agency, church organization, or private organization, etc.

source_descriptive_title
The title of the work, record, or item and, when appropriate, the title of the larger work or series of which it is a part.

source_filed_by_entry
This entry is to provide a short title used for sorting, filing, and retrieving source records.

source_publication_facts
When and where the record was created. For published works, this includes information such as the city of publication, name of the publisher, and year of publication.

text_from_source
A verbatim copy of any description contained within the source. This indicates notes or text that are actually contained in the source document, not the submitter's opinion about the source.

source_repository_citations
A list of SOURCE_REPOSITORY_CITATION() objects.

user_reference_number
A user-defined number or text that the submitter uses to identify this record.

user_reference_type
A user-defined definition of the user_reference_number.

automated_record_id
A unique record identification number assigned to the record by the source system. This number is intended to serve as a more sure means of identification of a record for reconciling differences in data between two interfacing systems.

date_changed
A CHANGE_DATE() object giving the time this record was last modified. If not provided, the current date is used.

notes
A list of NOTE_STRUCTURE() objects.

multimedia_links
A list of MULTIMEDIA_LINK() objects

Value

A tidy tibble containing a SOURCE_RECORD part of a GEDCOM file.