Activate a Submitter record

This allows a mechanism to easily edit particular records. Either the submitter name can be provided (can be a partial name) or the xref can be provided explicitly. Providing the name instead of the xref can make your code more readable, provided that the name returns a unique record!

activate_submitter_record(
  gedcom,
  submitter_name = character(),
  xref = character()
)

Arguments

gedcom
A tidygedcom object

submitter_name
The name of the submitter whose record should be activated

xref
The xref of the Submitter record to activate

Value

The same tidygedcom object with "active_record" attribute set to the specific Submitter record to allow easy editing.