Add a Family Group record to a tidygedcom object

This function adds a record containing details of a family group.

add_family_group(
  gedcom,
  husband = character(),
  wife = character(),
  children = character(),
  child_linkage_types = rep("birth", length(children)),
  number_of_children = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  family_notes = character(),
  multimedia_links = character()
)

Arguments

gedcom
A tidygedcom object.

husband
A character string identifying the husband of this family. This can either be an xref or a regular expression to match to an individual name.

wife
A character string identifying the wife of this family. This can either be an xref or a regular expression to match to an individual name.

children
A character vector of other Individual records that are children of this family. Children can either be referenced by an xref or by a regular expression to match to an individual name.

child_linkage_types
Codes used to indicate the child to family relationships. If defined, this must be a character vector the same size as children. Values must be one of: "birth" (default), "adopted", or "foster".

number_of_children
The reported number of children known to belong to this family, regardless of whether the associated children are represented here.

user_reference_number
A user-defined number or text that the submitter uses to identify this record. See the Gedcom 5.5.5 Specification for more details.

user_reference_type
A user-defined definition of the user_reference_number.

automated_record_id
A unique record identification number assigned to the record by the source system.

family_notes
A character vector of notes accompanying this Family group record. These could be xrefs to existing Note records.

multimedia_links
A character vector of multimedia file references accompanying this Family group record. These could be xrefs to existing Multimedia records.

Value

An updated tidygedcom object including the Family group record.

Details

This function will automatically assign a unique xref for this record. Most users will only need to use the husband, wife, children, and family_notes parameters (and of course gedcom).

If you need to add further information about this family (e.g. events), use the add_family_event_*() functions.

The function will automatically add links to this family to the respective Individual records of the wife, husband, and children.