Add a Multimedia record to a tidygedcom object

Add a Multimedia record to a tidygedcom object

add_multimedia(
  gedcom,
  file_reference,
  format,
  source_media = character(),
  title = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  multimedia_notes = character()
)

Arguments

gedcom
A tidygedcom object.

file_reference
A character vector of references for the files, typically a filepath or URL.

format
A character vector indicating the format of each multimedia file. Currently limited to one of: "AAC", "AVI", "BMP", "ePub", "FLAC", "GIF", "JPEG", "JPG", "MKV", "mobi", "MP3", "PCX", "PDF", "PNG", "TIFF", "TIF", "WAV". A format must be defined for each file reference given.

source_media
A code that indicates the type of material in which the referenced source is stored. Must be one of: audio, book, card, electronic, fiche, film, magazine, manuscript, map, newspaper, photo, tombstone, video. If this is defined, it must be a character vector the same size as file_reference and format.

title
The title of the multimedia file(s). If this is defined, it must be a character vector the same size as file_reference and format.

user_reference_number
A user-defined number or text that the submitter uses to identify this record. See the Gedcom 5.5.5 Specification for more details.

user_reference_type
A user-defined definition of the user_reference_number.

automated_record_id
A unique record identification number assigned to the record by the source system.

multimedia_notes
A character vector of notes accompanying this Multimedia record. These could be xrefs to existing Note records.

Value

An updated tidygedcom object including the Multimedia record.

Details

This function will automatically assign a unique xref for this record. Most users will only need to use the file_reference, format, source_media, title, and multimedia_notes parameters (and of course gedcom).