Extract a particular value from a tidygedcom object

Extract a particular value from a tidygedcom object

gedcom_value(gedcom, record_xref, tag, level, after_tag = NULL)

Arguments

gedcom
A tidygedcom object

record_xref
The xref of the record in which the value may exist

tag
The tag associated with the value

level
The level number of the value

after_tag
Whether the tag should be subordinate to this parent tag

Value

The particular value fitting the criteria of the input arguments. If not value is found, an empty string is returned. The function will automatically combine values split over CONT and CONC tags.